+++ src/contracts/readers/MarginReader.sol
@@ -6,15 +6,15 @@ import "../interfaces/Structs.sol";
 import "../libraries/MarginDirective.sol";
 
 contract MarginReader {
-
     // for this version of the sub account margin function, we make the assumption that we have
     // already computed a subAccount's potential spread positions
-    function getSubAccountMargin(
-        bool isInitial,
-        Structs.UserAndSystemState memory u
-    ) external pure returns (int256 health) {
+    function getSubAccountMargin(bool isInitial, Structs.UserAndSystemState memory u)
+        external
+        pure
+        returns (int256 health)
+    {
         // first we handle spots, this should have already had spreads deducted
-        for (uint i = 0; i < u.spots.length; i++) {
+        for (uint256 i = 0; i < u.spots.length; i++) {
             Structs.SpotPosition memory spot = u.spots[i];
             if (spot.spotRiskWeights.maintenanceLongWeight == 0) {
                 // in this case the spot asset is not valid collateral and contributes no health
@@ -22,9 +22,7 @@ contract MarginReader {
             }
             // check if the spot asset is core collateral, if so handle for it
             if (spot.spotAsset == u.coreCollateralAddress) {
-                health +=
-                    int256(spot.spotBalance) -
-                    int256(u.coreCollateralDebt);
+                health += int256(spot.spotBalance) - int256(u.coreCollateralDebt);
                 continue;
             }
             health += MarginDirective._calculateSpotHealth(
@@ -36,12 +34,10 @@ contract MarginReader {
             );
         }
         // second we handle spreads
-        for (uint i = 0; i < u.spreads.length; i++) {
+        for (uint256 i = 0; i < u.spreads.length; i++) {
             Structs.SpreadPosition memory spread = u.spreads[i];
             health += MarginDirective._calculateSpreadHealth(
-                isInitial
-                    ? spread.spreadPenalty.initial
-                    : spread.spreadPenalty.maintenance,
+                isInitial ? spread.spreadPenalty.initial : spread.spreadPenalty.maintenance,
                 spread.spreadQuantity,
                 spread.spotPrice,
                 spread.perpPrice,
@@ -51,7 +47,7 @@ contract MarginReader {
             );
         }
         // third we handle perps, this should have already had spreads deducted
-        for (uint i = 0; i < u.perps.length; i++) {
+        for (uint256 i = 0; i < u.perps.length; i++) {
             Structs.PerpPosition memory perp = u.perps[i];
             health += MarginDirective.getPerpMarginHealth(
                 isInitial,
